@comfyanonymous 2024-03-20
This should be fixed properly with: https://github.com/comfyanonymous/ComfyUI/commit/c18a203a8abdd0fce24743a838fe0d0400d8ff09